Uploaded Program 2,
Around 50 water samples collected from different regions of Agra were found to be contaminated with Cr6+ ions. Convert the normality of the samples taken from the 
user to ppm for all samples. Write a Java program for the above problem using arrays to take the input of data’s and to print the result.